+++ src/day2.h
@@ -39,7 +39,7 @@ namespace AOC23
         public:
             Day2(std::string fname);
             virtual int Part1();
-            virtual int Part2();
+            virtual unsigned long long Part2();
     };
 
 }
